Fix CRI Proxy hang
Due to wrong handling of channels, upon virtlet restart CRI Proxy
hanged on attempt to write second value to a one-time channel
which caused corresponding goroutine to hang. But since it was
done in a locked context, all subsequent kubelet API requests
involving virtlet queries (e.g. ListPodSandboxes) hang. Thus
k8s reported virtlet as not started (even though it was successfully
restarted), it's functionality was not available and the whole node
was marked as lost after some time.

Also, since this fix is prerequisite for #480, it contains
a fix to the e2e framework, which was also breaking the test in #480